fix(CommandContext): move enum to ts file
