+++ src/structures/CommandContext.ts
@@ -1,7 +1,7 @@
 /* eslint-disable @typescript-eslint/no-unnecessary-condition */
 import { ButtonInteraction, Collection, CommandInteraction, CommandInteractionOptionResolver, ContextMenuInteraction, GuildMember, Interaction, InteractionReplyOptions, Message, MessageOptions, MessagePayload, SelectMenuInteraction, TextBasedChannels, User } from "discord.js";
-import { InteractionTypes, MessageComponentTypes } from "discord.js/typings/enums";
 import { MessageInteractionAction } from "../typings";
+import { InteractionTypes, MessageComponentTypes } from "../typings/enum";
 
 
 export class CommandContext {
