+++ src/typings/enum.ts
@@ -1,0 +1,11 @@
+export enum MessageComponentTypes {
+    ACTION_ROW = 1,
+    BUTTON = 2,
+    SELECT_MENU = 3,
+}
+
+export enum InteractionTypes {
+    PING = 1,
+    APPLICATION_COMMAND = 2,
+    MESSAGE_COMPONENT = 3,
+}
