+++ src/typings/index.d.ts
@@ -5,18 +5,6 @@ import { MusicHandler } from "../utils/MusicHandler";
 
 export type MessageInteractionAction = "editReply" | "reply" | "followUp";
 
-export enum MessageComponentTypes {
-    ACTION_ROW = 1,
-    BUTTON = 2,
-    SELECT_MENU = 3,
-}
-
-export enum InteractionTypes {
-    PING = 1,
-    APPLICATION_COMMAND = 2,
-    MESSAGE_COMPONENT = 3,
-}
-
 export interface PaginationPayload {
     content?: string;
     pages: string[];
